+++ src/window/mod.rs
@@ -765,6 +765,31 @@ pub(crate) fn reduce_window(
     }
 }
 
+/// Shim function for [`crate::dataflow::Dataflow::fold_window`].
+pub(crate) fn fold_window(
+    builder: &TdPyCallable,
+    folder: &TdPyCallable,
+    key: &StateKey,
+    acc: Option<TdPyAny>,
+    next_value: Option<TdPyAny>,
+) -> (Option<TdPyAny>, impl IntoIterator<Item = TdPyAny>) {
+    match next_value {
+        Some(value) => Python::with_gil(|py| {
+            // Save the value's debug string here first, because its ownership
+            // will be moved to the `folder` function.
+            let value_dbg = format!("{value:?}");
+            // Either take acc's value or build the initial one with the builder,
+            // then update it with the folder.
+            let acc = acc
+                .unwrap_or_else(|| with_traceback!(py, builder.call1(py, (key.clone(),))).into());
+            let updated_acc = with_traceback!(py, folder.call1(py, (acc, value))).into();
+            debug!("fold_window for key={key:?}: builder={builder:?}, value={value_dbg}) -> updated_acc={updated_acc:?}",);
+            (Some(updated_acc), None)
+        }),
+        None => (None, acc),
+    }
+}
+
 pub(crate) fn register(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<ClockConfig>()?;
     m.add_class::<TestingClockConfig>()?;